My_AI_Advisor/requirements.txt
streamlit
pandas
numpy
yfinance>=0.2.40
pykrx
matplotlib
plotly
supabase
python-dateutil
pytz
langchain-community
langchain-core
langchain-google-genai
tavily-python
python-dotenv
